Move add_auth_to_remote to github module
This makes it more reusable across the app.
